requirements.txt
selenium==4.32.0
webdriver-manager==4.0.1
requests==2.31.0
python-dotenv==1.0.1
undetected-chromedriver==3.5.5 